Moved the bulk label code to be in its own file.
